moving self.metadata_sort into a concern as it's shared logic b/w models
